Move constants file down into _actions directory
That seems to be the only place they are used.
