+++ src/stratis_cli/_actions/_physical.py
@@ -21,11 +21,11 @@ from __future__ import print_function
 from stratisd_client_dbus import Pool
 from stratisd_client_dbus import get_object
 
-from .._constants import TOP_OBJECT
-
 from .._errors import StratisCliRuntimeError
 from .._errors import StratisCliUnimplementedError
 from .._stratisd_constants import StratisdErrors
+
+from ._constants import TOP_OBJECT
 
 from ._misc import GetObjectPath
