+++ tests/integration/test_misc.py
@@ -22,7 +22,7 @@ from stratisd_client_dbus import get_object
 
 from stratis_cli._actions._misc import GetObjectPath
 
-from stratis_cli._constants import TOP_OBJECT
+from stratis_cli._actions._constants import TOP_OBJECT
 from stratis_cli._errors import StratisCliDbusLookupError
 
 from ._misc import _device_list
